requirements.txt
bitsandbytes
transformers
peft
accelerate
trl
pandas
Pillow
wandb
sentencepiece
evaluate
tensorboardX
torch==2.4.1